requirements.txt


google-genai
pyyaml
requests
feedparser
beautifulsoup4
praw
